관리자 기능 이전
