+++ src/main/java/com/mapshot/api/presentation/notice/NoticeController.java
@@ -6,7 +6,6 @@ import com.mapshot.api.infra.auth.annotation.PreAuth;
 import com.mapshot.api.infra.auth.enums.Accessible;
 import com.mapshot.api.presentation.notice.model.NoticeDetailResponse;
 import com.mapshot.api.presentation.notice.model.NoticeListResponse;
-import com.mapshot.api.presentation.notice.model.NoticeRegistrationRequest;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.PositiveOrZero;
 import lombok.RequiredArgsConstructor;
@@ -43,31 +42,5 @@ public class NoticeController {
         NoticeDetailResponse noticeDetailResponse = noticeService.getSinglePost(postNumber);
 
         return ResponseEntity.ok(noticeDetailResponse);
-    }
-
-    @PreAuth(Accessible.ADMIN)
-    @PostMapping("/register")
-    public ResponseEntity<Void> registerNotice(@RequestBody NoticeRegistrationRequest noticeRegistrationRequest) {
-        noticeService.save(noticeRegistrationRequest);
-
-        return ResponseEntity.ok().build();
-    }
-
-    @PreAuth(Accessible.ADMIN)
-    @GetMapping("/delete/{noticeNumber}")
-    public ResponseEntity<Void> deleteNotice(@PositiveOrZero @PathVariable(value = "noticeNumber") long noticeNumber) {
-        noticeService.delete(noticeNumber);
-
-        return ResponseEntity.ok().build();
-    }
-
-
-    @PreAuth(Accessible.ADMIN)
-    @PostMapping("/modify/{noticeNumber}")
-    public ResponseEntity<Void> modifyNotice(@PositiveOrZero @PathVariable(value = "noticeNumber") long noticeNumber,
-                                             @RequestBody NoticeRegistrationRequest noticeRegistrationRequest) {
-        noticeService.modify(noticeNumber, noticeRegistrationRequest);
-
-        return ResponseEntity.ok().build();
     }
 }
